Refactore ticket controller
