+++ ws/src/main/scala/de/thm/ii/fbs/controller/practiceroom/ClassroomController.scala
@@ -136,171 +136,6 @@ class ClassroomController {
     }
   }
 
-  // Tickets
-  Tickets.onCreate((ticket) => {
-    smt.convertAndSend("/topic/classroom/" + ticket.courseId + "/ticket/create", ticketToJson(ticket))
-  })
-  Tickets.onUpdate((ticket) => {
-    smt.convertAndSend("/topic/classroom/" + ticket.courseId + "/ticket/update", ticketToJson(ticket))
-  })
-  Tickets.onRemove((ticket) => {
-    smt.convertAndSend("/topic/classroom/" + ticket.courseId + "/ticket/remove", ticketToJson(ticket))
-  })
-
-  /**
-    * Returns all tickets for a course
-    * @param m Composed ticket message.
-    * @param headerAccessor Header information
-    */
-  @MessageMapping(value = Array("/classroom/tickets"))
-  def getAllTickets(@Payload m: JsonNode, headerAccessor: SimpMessageHeaderAccessor): Unit = {
-    val localUserOpt = this.userService.loadCourseUserFromDB(headerAccessor.getUser.getName, m.retrive(courseIdLiteral).asInt().get);
-    val globalUserOpt = this.userService.find(headerAccessor.getUser.getName)
-    var userOpt = if (globalUserOpt.get.isAtLeastInRole(Role.TUTOR)) globalUserOpt else localUserOpt
-    localUserOpt match {
-      case Some(v) => userOpt = localUserOpt;
-      case None => userOpt = globalUserOpt;
-    }
-
-    val courseIdAndUser = for {
-      user <- userOpt
-      courseId <- m.retrive(courseIdLiteral).asInt()
-    } yield (courseId, user)
-
-    courseIdAndUser match {
-      case Some(v) => {
-        val (courseId, user) = v
-        if (!user.isAtLeastInRole(Role.TUTOR)) {
-          val response = Tickets.get(courseId)
-            .sortWith((a, b) => a.timestamp < b.timestamp)
-            .zipWithIndex.map { case (t: Ticket, i: Int) => {t.queuePosition = i + 1;t}}
-            .map { case(t: Ticket) => ticketToJson(t)}
-            .filter(t => t.getJSONObject("creator").get("username").toString == user.getName())
-            .foldLeft(new JSONArray())((a, t) => a.put(t))
-            .toString
-          smt.convertAndSendToUser(user.getName(), "/classroom/tickets", response)
-        } else {
-          val response = Tickets.get(courseId)
-            .sortWith((a, b) => a.timestamp < b.timestamp)
-            .zipWithIndex.map { case (t: Ticket, i: Int) => {t.queuePosition = i + 1;t}}
-            .map { case(t: Ticket) => ticketToJson(t)}
-            .foldLeft(new JSONArray())((a, t) => a.put(t))
-            .toString
-          smt.convertAndSendToUser(user.getName(), "/classroom/tickets", response)
-        }
-      }
-      case None => throw new MessagingException("Invalid content " + m)
-    }
-  }
-
-  /**
-    * Handles the creation of tickets
-    * @param m Composed ticket message.
-    * @param headerAccessor Header information
-    */
-  @MessageMapping(value = Array("/classroom/ticket/create"))
-  def createTicket(@Payload m: JsonNode, headerAccessor: SimpMessageHeaderAccessor): Unit = {
-    val title = m.retrive("title").toString.trim;
-    val desc = m.retrive("desc").toString.trim;
-    val priority = (m.retrive("priority").asInt()).asInstanceOf[Int];
-    if (title.isBlank){
-      throw new Exception("Title can not be empty");
-    }
-    if (desc.isBlank) {
-      throw new Exception("Description can not be empty")
-    };
-    if (priority > 0 && priority <= 10){
-      throw  new Exception("Priority must be between 0 and 11")
-    }
-    val ticketOpt = for {
-      user <- this.userService.loadCourseUserFromDB(headerAccessor.getUser.getName, m.retrive(courseIdLiteral).asInt().get)
-      courseId <- m.retrive(courseIdLiteral).asInt()
-      title <- m.retrive("title").asText()
-      desc <- m.retrive("desc").asText()
-      status <- m.retrive("status").asText()
-      timestamp <- m.retrive("timestamp").asLong()
-      priority <- m.retrive("priority").asInt()
-    } yield Tickets.create(courseId, title, desc, status, user, user, timestamp, priority)
-
-    if (ticketOpt.isEmpty) {
-      throw new MessagingException("Invalid msg: " + m)
-    }
-  }
-
-  /**
-    * Handles the update of tickets
-    * @param m Composed ticket message.
-    * @param headerAccessor Header information
-    */
-  @MessageMapping(value = Array("/classroom/ticket/update"))
-  def updateTicket(@Payload m: JsonNode, headerAccessor: SimpMessageHeaderAccessor): Unit = {
-    val ticketAndUser = for {
-      user <- this.userService.loadCourseUserFromDB(headerAccessor.getUser.getName, m.retrive(courseIdLiteral).asInt().get)
-      id <- m.retrive("id").asText()
-      courseId <- m.retrive(courseIdLiteral).asInt()
-      title <- m.retrive("title").asText()
-      desc <- m.retrive("desc").asText()
-      creator <- m.retrive("creator").asObject()
-      assignee <- m.retrive("assignee").asObject()
-      creatorName <- creator.retrive("username").asText()
-      assigneeName <- assignee.retrive("username").asText()
-      creatorAsUser <- userService.find(creatorName)
-      assigneeAsUser <- userService.find(assigneeName)
-      status <- m.retrive("status").asText()
-      timestamp <- m.retrive("timestamp").asLong()
-      priority <- m.retrive("priority").asInt()
-    } yield (practiceroom.Ticket(courseId, title, desc, status, creatorAsUser, assigneeAsUser, timestamp, priority, id), user)
-
-    ticketAndUser match {
-      case Some(v) => {
-        val (ticket, user) = v
-        if (!(ticket.creator.username == user.username || user.isAtLeastInRole(Role.TUTOR))) {
-          throw new MessagingException("User is not allowed to edit this ticket")
-        } else {
-          Tickets.update(ticket)
-        }
-      }
-      case None => throw new MessagingException("Invalid content: " + m)
-    }
-  }
-
-  /**
-    * Handles the removal of tickets
-    * @param m Composed ticket message.
-    * @param headerAccessor Header information
-    */
-  @MessageMapping(value = Array("/classroom/ticket/remove"))
-  def removeTicket(@Payload m: JsonNode, headerAccessor: SimpMessageHeaderAccessor): Unit = {
-    val ticketAndUser = for {
-      user <- this.userService.loadCourseUserFromDB(headerAccessor.getUser.getName, m.retrive(courseIdLiteral).asInt().get)
-      id <- m.retrive("id").asText()
-      ticket <- Tickets.getTicket(id)
-    } yield (ticket, user)
-
-    ticketAndUser match {
-      case Some(v) =>
-        val (ticket, user) = v
-        if (!(ticket.creator.username == user.username || user.isAtLeastInRole(Role.TUTOR))) {
-          throw new MessagingException("User is not allowed to remove this ticket")
-        } else {
-          Tickets.remove(ticket)
-        }
-      case None => throw new MessagingException("Invalid message: " + m)
-    }
-  }
-
-  private def ticketToJson(ticket: Ticket): JSONObject = new JSONObject()
-    .put("id", ticket.id)
-    .put("title", ticket.title)
-    .put("desc", ticket.desc)
-    .put("creator", userToJson(ticket.creator))
-    .put("assignee", userToJson(ticket.assignee))
-    .put("priority", ticket.priority)
-    .put("status", ticket.status)
-    .put(courseIdLiteral, ticket.courseId)
-    .put("timestamp", ticket.timestamp)
-    .put("queuePosition", ticket.queuePosition)
-
   /**
     * Handles the removal of tickets
     * @param m Composed ticket message.
